[Breaking] Publish toolchain containers
Starting with this commit the toolchain container is published as part
of the image workflow. This way users of the nativelink toolchains can
use the latest toolchains without having to build the images themselves.

Due to security vulnerabilities in older versions of Bazel the entire
Bazel dependency has been removed from the toolchain container and is
now only used during image generation. This means that the toolchain
container can no longer be used as an "invocation" container.

Users that used the toolchain container as "invocation" container in CI
workflows should migrate to the official `bazel` images or roll their
own images.