[Breaking] Migrate to Bzlmod
The new builds bootstrap the crate_universe binary which increases
compatibility with various platforms.

New supported builds:

- Native Bazel on MacOS
- Nix-wrapped Cargo fon MacOS

In addition, a new `--config=linux_zig` was added which allows building
nativelink with an hermetic toolchain that doesn't require nix. This
toolchain runs lot slower than the LRE toolchains but should produce
faster binaries (as it's also LLVM 17 under the hood) while being
backwards compatible down to Ubuntu 18. This means that users can now
skip installing `g++` (but still need `gcc`) in Dockerfile-style
container images and instead use `--config=linux_zig`.

This commit is somewhat of a regression in terms of hermeticity and
reproducibility. The bzlmod migration moved the Cargo.Bazel.lock file to
an autogenerated MODULE.bazel.lock, but we have to gitignore that file
for now since rules_rust currently doesn't create the same lockfiles on
Linux and MacOS.